Remove header styles
