+++ Calling/ClientApp/src/components/styles/ChatScreen.styles.ts
@@ -12,4 +12,30 @@ const chatScreenBottomContainerStyle = mergeStyles({
   overflow: 'auto'
 });
 
-export { chatScreenContainerStyle, chatScreenBottomContainerStyle };
+const chatHeaderContainerStyle = mergeStyles({
+  width: '100%',
+  height: 'auto',
+  paddingLeft: '3.25rem',
+  paddingRight: '3.25rem',
+  marginTop: '2rem',
+  selectors: {
+    '@media (max-width: 65rem)': {
+      paddingLeft: '5%',
+      paddingRight: '5%'
+    }
+  },
+  borderBottom: '0.063rem solid #DDDDDD'
+});
+
+const topicNameLabelStyle = mergeStyles({
+  color: '#323130',
+  fontSize: '1rem', // 16px
+  fontWeight: 600,
+  marginRight: '0.125rem',
+  width: '31.25rem',
+  wordBreak: 'break-all',
+  overflowY: 'hidden',
+  paddingBottom: '0.5rem'
+});
+
+export { chatScreenContainerStyle, chatScreenBottomContainerStyle, chatHeaderContainerStyle, topicNameLabelStyle };
